data_manage/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-snappy>=0.6.1
redis>=5.0.0
itsdangerous>=2.1.2
email-validator>=2.0.0
